// src/app/interface/controlNutricional.interface.ts
enum MedicionPersona {
    longitud_edad_0_2 = 'longitud_edad_0_2agno.json',
    peso_edad_0_5 = 'peso_edad_0_5agno.json',
    peso_longitud_0_2 = 'peso_longitud_0_2agno.json',
    peso_talla_2_5 = 'peso_talla_2_5agno.json',
    talla_edad_2_5 = 'talla_edad_2_5agno.json'
}

export const MedicionHombre = {
    ...MedicionPersona,
    asset: 'assets/data/hombre'
};
export const MedicionMujer = {
    ...MedicionPersona,
    asset: 'assets/data/mujer'
};
export enum TipoCalculo {
    longitud,
    talla
}
export enum Genero {
    Hombre = 'hombre',
    Mujer = 'mujer'
}

interface Medida {
    '1': string;
    '2': string;
    '3': string;
    '-3': string;
    '-2': string;
    '-1': string;
    Mediana: string;
}
export interface MedidaEdad extends Medida {
    'Año:Mes': string;
    Meses: string;
}

// Generated by https://quicktype.io

export interface PesoMedida extends Medida {
    cm: string;
}
export interface PesoEdad extends MedidaEdad {}